shuffle things about a bit
