+++ test/test_accession_id.py
@@ -1,4 +1,4 @@
-from transferrer.make_zip import shoot_number_to_accession_id
+from transferrer.transfer import shoot_number_to_accession_id
 import pytest
